+++ help/csharp/HowToHandshakeUtils/typedefinitions/RGBADefinition.cs
@@ -16,7 +16,7 @@ namespace RCP.Types
         {
         }
 
-        public override Parameter CreateParameter(short id, IParameterManager manager) => new ValueParameter<Color>(id, manager, this);
+        public override Parameter CreateParameter(int id, IParameterManager manager) => new ValueParameter<Color>(id, manager, this);
 
         public override Color ReadValue(KaitaiStream input)
         {
